+++ src/Silverback.Integration.MQTT/Messaging/Configuration/Mqtt/MqttProducerEndpointConfiguration.cs
@@ -30,6 +30,13 @@ public sealed record MqttProducerEndpointConfiguration : ProducerEndpointConfigu
     /// </summary>
     public uint MessageExpiryInterval { get; init; }
 
+    /// <summary>
+    ///     Gets a value indicating whether the producer should ignore the <i>NoMatchingSubscribers</i> error. When set to <c>true</c>, the
+    ///     error will be logged as a warning but the producer will not throw an exception.
+    ///     The default is <c>false</c>.
+    /// </summary>
+    public bool IgnoreNoMatchingSubscribersError { get; init; }
+
     /// <inheritdoc cref="EndpointConfiguration.ValidateCore" />
     protected override void ValidateCore()
     {